+++ airflow/contrib/kubernetes/kubernetes_request_factory/pod_request_factory.py
@@ -1,60 +1,67 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
+from .kubernetes_request_factory import *
 import yaml
-from airflow.contrib.kubernetes.kubernetes_request_factory.kubernetes_request_factory \
-    import KubernetesRequestFactory
-
+from airflow import AirflowException
 
 class SimplePodRequestFactory(KubernetesRequestFactory):
     """
-    Request generator for a simple pod.
+        Request generator for a simple pod.
     """
     _yaml = """apiVersion: v1
 kind: Pod
 metadata:
-  name: name
+  name: job_id
+  labels:
+      type: airflow
 spec:
-  containers:
-    - name: base
-      image: airflow-slave:latest
-      command: ["/usr/local/airflow/entrypoint.sh", "/bin/bash sleep 25"]
-  restartPolicy: Never
-    """
-
-    def __init__(self):
-        pass
+      containers:
+      - name: base
+        image: image
+        command: ["/usr/entrypoint"]
+      restartPolicy: Never
+"""
 
     def create(self, pod):
-        # type: (Pod) -> dict
         req = yaml.load(self._yaml)
-        self.extract_name(pod, req)
-        self.extract_labels(pod, req)
-        self.extract_image(pod, req)
-        self.extract_image_pull_policy(pod, req)
-        self.extract_cmds(pod, req)
-        self.extract_args(pod, req)
-        self.extract_node_selector(pod, req)
-        self.extract_env_and_secrets(pod, req)
-        self.extract_volume_secrets(pod, req)
-        self.attach_volumes(pod, req)
-        self.attach_volume_mounts(pod, req)
-        self.extract_resources(pod, req)
-        self.extract_service_account_name(pod, req)
-        self.extract_init_containers(pod, req)
-        self.extract_image_pull_secrets(pod, req)
+        extract_name(pod, req)
+        extract_labels(pod, req)
+        extract_image(pod, req)
+        extract_cmds(pod, req)
+        if len(pod.node_selectors) > 0:
+            extract_node_selector(pod, req)
+        extract_secrets(pod, req)
+        extract_volume_secrets(pod, req)
         return req
+
+
+class ReturnValuePodRequestFactory(SimplePodRequestFactory):
+    """
+    Pod request factory with a PreStop hook to upload return value
+    to the system's etcd service.
+    :param kube_com_service_factory: Kubernetes Communication Service factory
+    :type kube_com_service_factory: () => KubernetesCommunicationService
+    """
+
+    def __init__(self, kube_com_service_factory, result_data_file):
+        super(ReturnValuePodRequestFactory, self).__init__()
+        self._kube_com_service_factory = kube_com_service_factory
+        self._result_data_file = result_data_file
+
+    def after_create(self, body, pod):
+        """
+            Augment the pod with hyper-parameterized specific logic
+            Adds a Kubernetes PreStop hook to upload the model training
+            metrics to the Kubernetes communication engine (probably
+            an etcd service running with airflow)
+        """
+        container = body['spec']['containers'][0]
+        pre_stop_hook = self._kube_com_service_factory() \
+            .pod_pre_stop_hook(self._result_data_file, pod.name)
+        # Pre-stop hook only works on containers that are deleted. If the container
+        # naturally exists there would be no pre-stop hook execution. Therefore we
+        # simulate the hook by wrapping the exe command inside a script
+        if "'" in ' '.join(container['command']):
+            raise AirflowException('Please do not include single quote '
+                                   'in your command for hyperparameterized pods')
+        cmd = ' '.join(["'" + c + "'" if " " in c else c for c in container['command']])
+        container['command'] = ['/bin/bash', '-c', "({}) ; ({})"
+            .format(cmd, pre_stop_hook)]
