+++ airflow/contrib/operators/k8s_pod_operator/__init__.py
@@ -1,0 +1,1 @@
+from .k8s_pod_operator import *
