+++ scripts/ci/requirements.txt
@@ -94,3 +94,4 @@ thrift_sasl
 unicodecsv
 zdesk
 kubernetes
+yaml
